+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Initial Release
+- Added support for Aquaculture fishing rods